Investigate flaky add new visualization test (#15506) (#15667)
* try to repro

* try 30 times

* some clean up and extra tests

Don’t see how this should help, but switch to using data-test-subj and
add a wait until loading line.

Still, I don’t see that it would actually help.

Also it took 50 test run to hit it once, so doubling the amount of
times the test suite runs to improve chances of hitting it.

* Throw the function in a retry

Unfortunately it’s probably not worth investigating why the click
sometimes fails, given how rare it is and that it’s only reproducible
in the test environment.

* Remove extra test loops

* remove skip and check multiple times again

* Fix issue with flaky field toggle selection

* forgot await

* More stabilizations

* Stabilize by disabling toast auto hides on all dashboard tests

* avoid all the retries to see if disabling toast auto hides was what was messing everything up

* remove comment that is no longer relevant

* Remove extra tests